examples directory cleanup - see ext
previously, we had the `.gitignore` configured to disregard _everything_ in the the `examples` directory and then we added 'negative' rules to include the files we want, this change inverts that logic s.t. we ignore only the output files that we don't want. This was done because there are some file extensions that are output as part of the examples that are also used elsewhere to test as input, which made the negative rules difficult and tedious.